requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.2
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==42.0.8
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==33.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jwt==1.3.1
limits==5.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
packaging==24.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
sentry-sdk==2.11.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.29.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2